🚚 Move tests to EU-DEMO directory
Can't be named the same as tests because pytest gets confused
