+++ .flake8
@@ -50,6 +50,31 @@ per-file-ignores =
         D409,
         # It can be useful to have class names in test names e.g., test_raises_SomeException
         N802,
+    examples/design/EU-DEMO/EUDEMO_tests/*:
+     	# Use of assert detected. The enclosed code will be removed when compiling
+	    # to optimised byte code. (ignored because of pytest)
+        S101,
+	    # Sometimes the tests just run code without using local variables
+        F841,
+        # We don't require documentation in tests
+        D100,
+        D101,
+        D102,
+        D103,
+        D104,
+        D105,
+        D202,
+        D205,
+        D210,
+        D300,
+        D301,
+        D401,
+        D403,
+        D404,
+        D406,
+        D409,
+        # It can be useful to have class names in test names e.g., test_raises_SomeException
+        N802,
     __init__.py:
         # Allow unused imports in __init__.py, so we can expose at parent module level
         F401,
