+++ examples/design/EU-DEMO/EUDEMO_tests/test_data/tf_coils_TripleArc_18.json
