+++ examples/design/EU-DEMO/EUDEMO_tests/test_equilibria.py
@@ -19,31 +19,27 @@
 # You should have received a copy of the GNU Lesser General Public
 # License along with bluemira; if not, see <https://www.gnu.org/licenses/>.
 
-from bluemira.base.config import Configuration
-from bluemira.builders.EUDEMO.maintenance import UpperPortOP
-from bluemira.geometry.face import BluemiraFace
-from bluemira.geometry.tools import make_polygon
-from bluemira.utilities.optimiser import Optimiser
+import numpy as np
+import pytest
+from EUDEMO_builders.equilibria import estimate_kappa95
 
 
-class TestUpperPortOP:
-    def test_dummy_blanket_port_opt(self):
-        params = Configuration()
-        bb = make_polygon(
-            {
-                "x": [5, 6, 6, 11, 11, 12, 12, 5],
-                "y": 0,
-                "z": [-5, -5, 5, 5, -5, -5, 6, 6],
-            },
-            closed=True,
-        )
-        bb = BluemiraFace(bb)
-        optimiser = Optimiser(
-            "SLSQP", opt_conditions={"max_eval": 1000, "ftol_rel": 1e-8}
-        )
+class TestKappaLaw:
+    """
+    As per the conclusions of the CREATE report 2L4NMJ
+    """
 
-        design_problem = UpperPortOP(params, optimiser, bb)
-
-        solution = design_problem.optimise()
-
-        assert design_problem.opt.check_constraints(solution)
+    @pytest.mark.parametrize(
+        "A, m_s, expected",
+        [
+            [3.6, 0.3, 1.58],
+            [3.1, 0.3, 1.68],
+            [2.6, 0.3, 1.73],
+            [3.6, 0, 1.66],
+            [3.1, 0, 1.77],
+            [2.6, 0, 1.80],
+        ],
+    )
+    def test_kappa(self, A, m_s, expected):
+        k95 = estimate_kappa95(A, m_s)
+        np.testing.assert_allclose(k95, expected, rtol=5e-3)
